sql: Comment and test types package
Rearrange and clean up methods in the package, including moving some
functionality from the package to other locations (e.g. information
schema related methods). Add comments to all exported structs and methods
in the types package. Add tests for all functionality in the package.

Fix compatibility issue in pg_type table, where oidvector and int2vector
had array oid types of 0.

Release note (sql change): Display correct array type oids for oidvector
and int2vector types in pg_type table, for better PG compatibility.

+++ pkg/acceptance/testdata/java/src/main/java/MainTest.java
@@ -174,8 +174,8 @@ public class MainTest extends CockroachDBTest {
         rs.next();
 
         Array ar = rs.getArray(1);
-        Long[] fs = (Long[]) ar.getArray();
-        Assert.assertArrayEquals(new Long[]{123L}, fs);
+        Integer[] fs = (Integer[]) ar.getArray();
+        Assert.assertArrayEquals(new Integer[]{123}, fs);
     }
 
     @Test
